migrations style fix
